requirements.txt
dotenv
langchain
langchain-groq
langchain-core
langgraph
langdetect
windows-curses
edge_tts
indic_transliteration
deep_translator
mutagen
python-vlc
SpeechRecognition
google-generativeai
Pillow
opencv-python
pyperclip
pyaudio
pandas
